register-service/requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.2
cycler==0.12.1
dotmap==1.3.30
exceptiongroup==1.3.1
fastapi==0.135.1
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.9.0
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.75.1
h11==0.16.0
h5py==3.14.0
idna==3.10
Jinja2==3.1.6
kagglehub==1.0.0
kagglesdk==0.1.15
keras==3.11.3
kiwisolver==1.4.9
libclang==18.1.1
lightning-utilities==0.15.3
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
multidict==6.7.1
mysql-connector-python==9.6.0
namex==0.1.0
networkx==3.4.2
numpy==2.2.6
onnx==1.20.1
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
platformdirs==4.5.0
propcache==0.4.1
protobuf==6.32.1
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.1
PyYAML==6.0.3
requests==2.32.5
rich==14.1.0
scipy==1.15.3
sentry-sdk==2.42.1
simsimd==6.5.3
six==1.17.0
smmap==5.0.2
starlette==0.52.1
stringzilla==4.2.3
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
torch==2.8.0
torch_fidelity==0.4.0
torchaudio==2.8.0
torchmetrics==1.8.2
torchvision==0.23.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.41.0
wandb==0.22.2
Werkzeug==3.1.3
wrapt==1.17.3
yarl==1.22.0
